[react=ui] organize stuff into utils folder
