+++ assets/play_react/js/utils/guid.js
